refactor : 패키지 이동
- controller/domain/view 분리
