+++ src/main/kotlin/racingcar/domain/ConditionProvider.kt
@@ -1,4 +1,4 @@
-package racingcar.controller
+package racingcar.domain
 
 interface ConditionProvider {
     fun nextCondition(): Int
