+++ src/test/kotlin/racingcar/domain/CarTest.kt
@@ -1,4 +1,4 @@
-package racingcar.controller
+package racingcar.domain
 
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.StringSpec
